# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cautai AI Search Engine API
  description: |
    Cautai is an AI-first search engine designed for both agents and humans. This API provides 
    comprehensive search capabilities, content composition, and citation generation services.
    
    ## Features
    - Advanced search with hybrid ranking (BM25 + semantic similarity)
    - AI-powered content composition and summarization
    - Academic citation generation in multiple formats
    - Multi-language support (English, Romanian)
    - Rate limiting and authentication
    - Real-time search result processing
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain tokens through the authentication endpoints.
    
    ## Rate Limiting
    - Search endpoints: 100 requests per minute per IP
    - Composition endpoints: 20 requests per minute per authenticated user
    - Citation endpoints: 50 requests per minute per authenticated user
    
  version: 1.0.0
  contact:
    name: Cautai Support
    url: https://cautai.ro/support
    email: support@cautai.ro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cautai.ro/v1
    description: Production server
  - url: https://staging-api.cautai.ro/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /auth/login:
    post:
      summary: User authentication
      description: Authenticate user and receive access token
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          description: User password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Unique request identifier for debugging
...